Apply changes from PR comments
