+++ app/src/main/java/com/example/footballmanager/ui/FootballManagerApp.kt
@@ -14,18 +14,18 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.tooling.preview.Preview
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.footballmanager.screens.HomeScreen
-import com.example.footballmanager.ui.theme.Header
 import com.example.footballmanager.screens.view_models.HomeViewModel
+import com.example.footballmanager.ui.theme.Header
 import com.example.footballmanager.ui.theme.background
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun FootballManagerApp() {
     val scrollBehavior = TopAppBarDefaults.enterAlwaysScrollBehavior()
-    val homeViewModel: HomeViewModel = viewModel()
+    val homeViewModel: HomeViewModel = hiltViewModel()
     val ctx = LocalContext.current
     LaunchedEffect(key1 = true) {
         homeViewModel.getFixturesData(ctx = ctx)
